offscreen-canvas/the-offscreen-canvas/offscreencanvas.commit.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="help" href="https://html.spec.whatwg.org/multipage/scripting.html#the-offscreen-2d-rendering-context">
<script>

function examPlaceholder(placeholder)
{
    var canvas = document.createElement('canvas');
    canvas.width = canvas.height = 10;
    var ctx = canvas.getContext('2d');
    ctx.drawImage(placeholder, 0, 0);
    _assertPixel(canvas, 5,5, 0,255,0,255, "5,5", "0,255,0,255");
}

test(function() {
    var placeholder = document.createElement('canvas');
    placeholder.width = placeholder.height = 10;
    var offscreenCanvas = placeholder.transferControlToOffscreen();
    var ctx = offscreenCanvas.getContext('2d');
    ctx.fillStyle = "#0f0";
    ctx.fillRect(0, 0, 10, 10);
    ctx.commit();
    // Set timeout acts as a sync barrier to allow commit to propagate
    setTimeout(function() {
        examPlaceholder(placeholder);
    }, 0);
}, "Test that commit behaves correctly.");

test(function() {
    var offscreenCanvas = new OffscreenCanvas(10, 10);
    var ctx = offscreenCanvas.getContext('2d');
    ctx.fillStyle = "#0f0";
    ctx.fillRect(0, 0, 10, 10);
    assert_throws("InvalidStateError", function() { ctx.commit(); });
}, "Test that call commit on an OffscreenCanvas that is not transferred from a HTMLCanvasElement throws exception.");

</script>
